Rename `BatchAllocation` to `BatchAllocations`
