+++ src/Groups/GroupDetails/GroupDetails.js
@@ -30,7 +30,7 @@ import {
   useModalToggle,
 } from '@folio/stripes-acq-components';
 
-import { BatchAllocationModal } from '../../common/components/BatchAllocation';
+import { BatchAllocationModal } from '../../common/components/BatchAllocations';
 import {
   DetailsEditAction,
   DetailsRemoveAction,
