+++ src/common/components/BatchAllocations/BatchAllocationView/BatchAllocationView.js
