+++ src/common/components/BatchAllocations/BatchAllocationView/index.js
