+++ src/common/components/BatchAllocations/constants.js
