+++ src/common/components/BatchAllocations/index.js
