+++ src/common/components/BatchAllocations/utils.js
