+++ src/common/components/index.js
@@ -1,3 +1,3 @@
 export { AllocationToolsMenuSection } from './AllocationToolsMenuSection';
 export { DownloadAllocationWorksheetModal } from './DownloadAllocationWorksheetModal';
-export * from './BatchAllocation';
+export * from './BatchAllocations';
